partially document /variables
