+++ bytecode/src/variables.rs
@@ -2,7 +2,6 @@ mod object;
 mod ops;
 mod primitive;
 mod primitive_shorthands;
-mod variable;
 
 pub use object::Object;
 pub use object::ObjectBuilder;
